Rename module path to gobl.it.sdi
