+++ addon/invoice_scenarios_test.go
@@ -3,7 +3,7 @@ package sdi_test
 import (
 	"testing"
 
-	sdi "github.com/invopop/gobl.fatturapa/addon"
+	sdi "github.com/invopop/gobl.it.sdi/addon"
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cal"
 	"github.com/invopop/gobl/cbc"
