+++ addon/tax_combo_test.go
@@ -3,7 +3,7 @@ package sdi_test
 import (
 	"testing"
 
-	sdi "github.com/invopop/gobl.fatturapa/addon"
+	sdi "github.com/invopop/gobl.it.sdi/addon"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/norm"
 	"github.com/invopop/gobl/num"
