+++ cmd/gobl.fatturapa/convert.go
@@ -9,7 +9,7 @@ import (
 	"path/filepath"
 
 	"github.com/invopop/gobl"
-	fatturapa "github.com/invopop/gobl.fatturapa"
+	fatturapa "github.com/invopop/gobl.it.sdi"
 	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/xmldsig"
 	"github.com/spf13/cobra"
