+++ items_parse_test.go
@@ -5,8 +5,8 @@ import (
 	"path/filepath"
 	"testing"
 
-	sdi "github.com/invopop/gobl.fatturapa/addon"
-	"github.com/invopop/gobl.fatturapa/test"
+	sdi "github.com/invopop/gobl.it.sdi/addon"
+	"github.com/invopop/gobl.it.sdi/test"
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/num"
