+++ parties_test.go
@@ -3,8 +3,8 @@ package fatturapa_test
 import (
 	"testing"
 
-	sdi "github.com/invopop/gobl.fatturapa/addon"
-	"github.com/invopop/gobl.fatturapa/test"
+	sdi "github.com/invopop/gobl.it.sdi/addon"
+	"github.com/invopop/gobl.it.sdi/test"
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/l10n"
