+++ transmission.go
@@ -2,7 +2,7 @@ package fatturapa
 
 import (
 	"github.com/invopop/gobl"
-	sdi "github.com/invopop/gobl.fatturapa/addon"
+	sdi "github.com/invopop/gobl.it.sdi/addon"
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/gobl/org"
