+++ transmission_test.go
@@ -3,7 +3,7 @@ package fatturapa_test
 import (
 	"testing"
 
-	"github.com/invopop/gobl.fatturapa/test"
+	"github.com/invopop/gobl.it.sdi/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
